speech-to-text-app/requirements.txt
Flask==2.3.3
whisper-openai==1.0.0  # Use a stable Whisper package
gunicorn==21.2.0      # For production-ready server